Move ensemble_wait to backend
